+++ changelog.md
@@ -3,7 +3,9 @@
 - Added the command `/ess convert EssentialsX`, allows converting the database from EssentialsX to zEssentials.
   Documentation: https://zessentials.groupez.dev/getting-started/convert#essentialsx
 - Fixed the scoreboard title not appearing
-- Added/homes alias for/home command for default configurations
+- Added `/homes` alias for `/home` command for default configurations
+- Added `/enchant` command
+- Added enchantments list with aliases
 
 # 1.0.0.1